requirements.txt
# requirements.txt

asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
distlib==0.3.9
Django==5.1.2
djangorestframework==3.15.2
dnspython==2.7.0
filelock==3.16.1
identify==2.6.1
idna==3.10
mongoengine==0.29.1
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
